% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data for multisc package}
\format{
A list that contains four elements: raw count matrix, sample-level metadata, cell-level metadata and leaves info generated from TreeCorTreat
data(example_data)
}
\usage{
data(example_data)
}
\description{
Processed single-cell RNA-seq data extracted from multiple COVID-19 studies. 20 samples (10 mild + 10 moderate) in total.
}
\keyword{datasets}
