% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_clean.R
\name{runPseudoAnalysis}
\alias{runPseudoAnalysis}
\title{RunPseudoAnalysis}
\usage{
runPseudoAnalysis(
  raw_count,
  cell_meta,
  sample_meta,
  leaves_info,
  features = "expr_and_prop",
  batch = NULL,
  pb_filter_pct = 0.9,
  pb_HVG = T,
  pb_qnorm = T,
  pb_batch_correction = F,
  pb_parallel = F,
  num_cores = 10,
  ctp_clr = F,
  ctp_adj_cov = NULL,
  ctp_batch_correction = F,
  sample_name = "sample",
  severity_name = "severity",
  severity_num_name = "sev.level",
  ptime_method = "cca",
  clustermethod = "louvain",
  cluster_id = NULL,
  reduce = F,
  clusternum = 2,
  startcluster = 1,
  near_nbr_num = 3,
  modelNames = "VVV",
  pc_scale = T,
  eval_method = "corr",
  pseudotime_name = "pseudotime",
  gene_mapping = F,
  deg_parallel = F,
  deg_num_cores = 10,
  save_deg = F,
  sample_embed_plot = F,
  result_path = "./result/"
)
}
\arguments{
\item{result_path}{}
}
\value{

}
\description{
RunPseudoAnalysis
}
