% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_clean.R
\name{runPseudoDEG}
\alias{runPseudoDEG}
\title{runPseudoDEG}
\usage{
runPseudoDEG(
  sample_meta,
  pb.ls,
  leaves_info,
  severity_name = "severity",
  sample_name = "sample",
  pseudotime_name = "pseudotime",
  gene_mapping = F,
  deg_parallel = F,
  deg_num_cores = 10
)
}
\arguments{
\item{deg_num_cores}{}
}
\value{

}
\description{
runPseudoDEG
}
