% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_clean.R
\name{saveResults}
\alias{saveResults}
\title{saveResults}
\usage{
saveResults(
  ptime_result = NULL,
  leaves_info = NULL,
  ptime_method = "cca",
  pseudotime_name = "pseudotime",
  sample_name = "sample",
  severity_name = "severity",
  gene_mapping = F,
  deg_parallel = F,
  deg_num_cores = 10,
  save_deg = F,
  sample_embed_plot = F,
  result_path = "./result/"
)
}
\arguments{
\item{result_path}{}
}
\value{

}
\description{
saveResults
}
